requirements.txt
python-dotenv
pymongo
loguru
requests